Cleaned up some docs and example code, also gave view models a method to apply styles to themselves.
