+++ super_editor/example/lib/demos/components/demo_unselectable_hr.dart
@@ -52,10 +52,8 @@ class _UnselectableHrDemoState extends State<UnselectableHrDemo> {
   Widget build(BuildContext context) {
     return SuperEditor(
       editor: _docEditor,
-      stylesheet: Stylesheet(
+      stylesheet: defaultStylesheet.copyWith(
         documentPadding: const EdgeInsets.symmetric(vertical: 56, horizontal: 24),
-        rules: defaultStylesheet.rules,
-        inlineTextStyler: defaultStylesheet.inlineTextStyler,
       ),
       // Add a new component builder that creates an unselectable
       // horizontal rule, instead of creating the usual selectable kind.
