+++ super_editor/example/lib/demos/demo_document_loses_focus.dart
@@ -35,10 +35,8 @@ class _LoseFocusDemoState extends State<LoseFocusDemo> {
           Expanded(
             child: SuperEditor(
               editor: _docEditor,
-              stylesheet: Stylesheet(
+              stylesheet: defaultStylesheet.copyWith(
                 documentPadding: const EdgeInsets.symmetric(vertical: 56, horizontal: 24),
-                rules: defaultStylesheet.rules,
-                inlineTextStyler: defaultStylesheet.inlineTextStyler,
               ),
             ),
           ),
