+++ super_editor/example/lib/demos/demo_markdown_serialization.dart
@@ -64,10 +64,8 @@ class _MarkdownSerializationDemoState extends State<MarkdownSerializationDemo> {
             child: SuperEditor(
               key: _docKey,
               editor: _docEditor,
-              stylesheet: Stylesheet(
+              stylesheet: defaultStylesheet.copyWith(
                 documentPadding: const EdgeInsets.symmetric(vertical: 56, horizontal: 24),
-                rules: defaultStylesheet.rules,
-                inlineTextStyler: defaultStylesheet.inlineTextStyler,
               ),
             ),
           ),
