+++ super_editor/example/lib/demos/demo_paragraphs.dart
@@ -28,10 +28,8 @@ class _ParagraphsDemoState extends State<ParagraphsDemo> {
   Widget build(BuildContext context) {
     return SuperEditor(
       editor: _docEditor,
-      stylesheet: Stylesheet(
+      stylesheet: defaultStylesheet.copyWith(
         documentPadding: const EdgeInsets.symmetric(vertical: 56, horizontal: 24),
-        rules: defaultStylesheet.rules,
-        inlineTextStyler: defaultStylesheet.inlineTextStyler,
       ),
     );
   }
