+++ super_editor/example/lib/demos/sliver_example_editor.dart
@@ -57,10 +57,8 @@ class _SliverExampleEditorState extends State<SliverExampleEditor> {
         SliverToBoxAdapter(
           child: SuperEditor(
             editor: _docEditor,
-            stylesheet: Stylesheet(
+            stylesheet: defaultStylesheet.copyWith(
               documentPadding: const EdgeInsets.symmetric(vertical: 56, horizontal: 24),
-              rules: defaultStylesheet.rules,
-              inlineTextStyler: defaultStylesheet.inlineTextStyler,
             ),
           ),
         ),
