+++ super_editor/example/lib/marketing_video/main_marketing_video.dart
@@ -198,16 +198,15 @@ class _MarketingVideoState extends State<MarketingVideo> {
           documentLayoutKey: _docLayoutKey,
           editor: _editor,
           composer: _composer,
-          stylesheet: Stylesheet(
+          stylesheet: defaultStylesheet.copyWith(
             documentPadding: const EdgeInsets.all(16),
-            rules: defaultStylesheet.rules
-              ..add(
-                StyleRule(
-                    const BlockSelector.all(),
-                    (doc, node) => {
-                          "padding": const CascadingPadding.all(0.0),
-                        }),
-              ),
+            addRulesAfter: [
+              StyleRule(
+                  const BlockSelector.all(),
+                  (doc, node) => {
+                        "padding": const CascadingPadding.all(0.0),
+                      }),
+            ],
             inlineTextStyler: (attributions, style) => _textStyleBuilder(attributions),
           ),
         ),
