+++ super_editor/lib/src/default_editor/layout_single_column/layout_single_column.dart
@@ -3,3 +3,6 @@
 
 export '_layout.dart';
 export '_presenter.dart';
+export '_styler_per_component.dart';
+export '_styler_shylesheet.dart';
+export '_styler_user_selection.dart';
